working local tests

+++ test_scripts/train_corridor.sh
@@ -3,31 +3,47 @@
 #SBATCH --cpus-per-task=2
 #SBATCH --time=0-24:00:00  # DD-HH:MM:SS
 
-
+CUR_DIR=$(pwd)
 SLURM_TMPDIR='./output'
 
 if [ ! -d "$SLURM_TMPDIR" ]
 then
+  echo 'creating slur tmpdir'
   mkdir $SLURM_TMPDIR
 fi
 
 # Setup venv
 cd $SLURM_TMPDIR
-virtualenv --no-download env
 source ~/.venv_ift725/bin/activate
 pip install --no-index -r ~/repos/CuriousNetwork_fixes/requirements.txt
 
 # Copy code and data
-cd $SLURM_TMPDIR
-cp -r ~/repos/CuriousNetwork/nn/ .
-cd nn/
-cp ~/repos/CuriousNetwork/dataset.tar .
-tar -xf dataset.tar
-cp ~/CuriousNetwork_fixes/barbar.py .
+if [ ! -d "./nn" ]
+then
+  echo 'copying nn to slurm tmpdir'
+  cp -r ~/repos/CuriousNetwork/nn/ .
+fi
 
-cd $SLURM_TMPDIR/nn/
+cd nn/
+
+if [ ! -e "./dataset.tar" ]
+then
+  echo 'copying dataset to slurm tmpdir'
+  cp ~/repos/CuriousNetwork/dataset.tar .
+fi
+
+if [ ! -d "./dataset" ]
+then
+  echo 'extracting dataset'
+  tar -xf dataset.tar
+fi
+
+cp ~/repos/CuriousNetwork_fixes/barbar.py .
 
 # Start the training
-python train.py --train_path $SLURM_TMPDIR/nn/dataset/jpg/corridor_train \
-                --val_path $SLURM_TMPDIR/nn/dataset/jpg/corridor_val \
-                --test_path $SLURM_TMPDIR/nn/dataset/jpg/corridor_test "$@"
+echo 'initiating training'
+python train.py --train_path /home/ggs22/repos/CuriousNetwork/test_scripts/output/nn/dataset/jpg/corridor_train \
+                --val_path /home/ggs22/repos/CuriousNetwork/test_scripts/output/nn/dataset/jpg/corridor_val \
+                --test_path /home/ggs22/repos/CuriousNetwork/test_scripts/output/nn/dataset/jpg/corridor_test "$@"
+
+cd $CUR_DIR
